Added Parameters. Changed name.
I changed the class name from OpenSim::ExponentialSpringForce, which might have been confused with SimTK::ExponentialSpringForce, to OpenSim::ExponentialContact.

I also added the topology-stage parameters.  On the SimTK side, they are managed by SimTK::ExponentialSpringParameters.  I tried not to reproduce functionality on the OpenSim side, but to make use of SimTK::ExponentialSpringParameters as much as possible.  OpenSim::ExponentialContact has its own "Parameters" class that encapsulates a SimTK::ExponentialSpringParameters instance.  The main purpose of OpenSim::Parameters is to establish and manage all of the OpenSim Properties that correspond to the member variables in SimTK::ExponentialSpringParameters.