Made naming more consistent.
ExponentialSpringForce -> ExponentialContact, and the abbreviation at the end of variable names when from "ES" to "EC".

I also re-organized some of the model initialization so that static model testing could precede the simulation.  Basically, a model needs to be completely put together before meaningful tests can be conducted, AND the state needs to be initialized (and the SimTK::System needs to be realized at Stage::Topology) just prior to integration.