ExponentialContact:: added methods related to Discrete States.
I added several methods to support exposing the discrete states of class ExponentialContact to the OpenSim::Component API.

I added the following methods to establish the names that class Component will use for the Discrete Variable names.  These methods allow one to avoid placing quoted strings a bunch of places in the code.  They are mainly intended for construction / initialization related methods.
getMuStaticDiscreteStateName()
getMuKineticDiscreteStateName()
getSlidingDiscreteStateName().

I added ExponentialContact::extendRealizeTopology().  This method initializes the index and Subsystem for each Discrete Variable.  This initialization needs to occur in ExponentialContact because the Discrete Variables are no longer allocated in Component::extendRealizeTopology() (see comment below).  Note that the Subsystem needs to be initialized because the Discrete Variables that belong to ExponentialContact are allocated from a different SimTK::Subsystem than is normally used by class Component.

Finally, in ExponentialContact::extendAddToSystem(), note that the calls to addDiscreteVariable() include the 'allocate' flag with a value of 'false'.  This prevents the double allocation of the discrete variables in Component::extendRealizeTopology().