Removed trailing whitespace in Component.cpp
There are no substantive modifications in this commit.  I've just removed trailing white spaces.

I'd like to merge this commit to OpenSim Core so that the substantive modifications that I'm making to class Component in branch exponential_springs won't be camouflaged by all the non-substantive whitespace changes when it comes times to review the exponential_springs branch code.