[WIP] Modifications to support serializing Discrete Variables.
The code is not working in this current commit.  Continuing from the current modifications, I am going to experiment with several approaches to support serializing discrete Variaibles.  Rather than making changes in the current code, I'm going to branch.  When I find an approach that works, I will merge back.