Removed wip methods for traversing to a DiscreteVariable.
These methods were a work in progress [wip] and were never actually functioning.  They will only be needed if the decision is to go forward with serializing/deserializing discrete states.  If so, then it's probably best to start from a clean slate on these anyway.

This commit is also accompanied by a few minor typo corrections.