New TEST_CASE: Component Interface State Trajectories
A simple system possessing 4 state variables and 2 discrete state variables is constructed.  Based on an artificial simulation, a state trajectory (SimTK::Array_<SimTK::State> is constructed.

Based on this state trajectory, Component::getStateVariableTrajectory() and Component::setStateVariableTrajectory() are used to walk through the essential steps of a serialization and deserialization.

All checks in this test case are passing.