Update PiecewiseLinearFunction.cpp
- Added a SimTK_ASSERT() that traps when calcValue() is called with less than 2 data points.

At least 2 data points are required for a valid interpolation. Without this assert, it is possible that uninitialized or out-of-bounds elements of the coefficient array, _b, would be used, leading to a sporadically invalid simulation or analysis.

A performance benefit of using SimTK_ASSERT() is that the assert is performed in a Debug build but not in a Release build.