Update PiecewiseLinearFunction.cpp
- Removed the SimTK_ASSERT() I added in the last commit.
- Altered calcCoefficients() to size and initialize the coefficient array, _b, when the number of function points is 1.

It turns out that class PiecewiseLinearFunction is able to perform a valid calcValue() when there is only 1 function point. Throwing an exception when size < 2, therefore, is not a valid thing to do. So, I removed the SimTK_ASSERT() that I added in the previous commit.

The issue was that, unless the number of function points was 2 or more, the array of coefficients array, _b, was not being sized and initialized. This issue is addressed by the code changes in lines 417 to 425