Accept paths as Path objects (not only strings)
Python pathlib paths are much cleaner and easier to use than strings, and it is the recommended Python approach for manipulating paths since Python 3.6 (introduced in Python 3.4). When using OpenSim in other libraries that work with Path objects, it is a bit of a pain and a sourve of errors to have to convert the path into a string.

The PR lets the OpenSim Python bindings accept paths both as strings and Path objects. 

Please note that I spent a full day trying to build it from source without success (I'm definitely lacking practice for this), and I feel like I need to let it go for now.  

I'm fairly confident that it would work, but I did not test it. At the very least, someone should try to build it and test this, at the very least:
```python
import opensim as osim
from pathlib import Path
model = osim.Model("arm26.osim")
model = osim.Model(Path("arm26.osim"))
```