Add CantileverFreeBeamJoint
This joint type provides a lightweight way for modeling flexible structures (e.g., the bending of the spinal column). It has only three rotational generalized cooordinates, but the parent frame translates relative to the child frame as a specified function of the coordinates according to relationships derived from beam bending theory for a cantilever-free beam.